fix an issue compiling the GUI on Ubuntu > Trusty

+++ src/qt/proposaltablemodel.cpp
@@ -123,11 +123,11 @@ QVariant ProposalTableModel::data(const QModelIndex &index, int role) const
         case Proposal:
             return rec->name;
         case YesVotes:
-            return rec->yesVotes;
+            return (long long) rec->yesVotes;
         case NoVotes:
-            return rec->noVotes;
+            return (long long) rec->noVotes;
         case AbsoluteYesVotes:
-            return rec->absoluteYesVotes;
+            return (long long) rec->absoluteYesVotes;
         case StartDate:
             return (QDateTime::fromTime_t((qint32)rec->start_epoch)).date().toString(Qt::SystemLocaleLongDate);
         case EndDate:
@@ -148,15 +148,15 @@ QVariant ProposalTableModel::data(const QModelIndex &index, int role) const
         case EndDate:
             return rec->end_epoch;
         case YesVotes:
-            return rec->yesVotes;
+            return (long long) rec->yesVotes;
         case NoVotes:
-            return rec->noVotes;
+            return (long long) rec->noVotes;
         case AbsoluteYesVotes:
-            return rec->absoluteYesVotes;
+            return (long long) rec->absoluteYesVotes;
         case Amount:
-            return qint64(rec->amount);
+            return qint64((long long) rec->amount);
         case Percentage:
-            return rec->percentage;
+            return (long long) rec->percentage;
         }
         break;
     case Qt::TextAlignmentRole:
@@ -175,19 +175,19 @@ QVariant ProposalTableModel::data(const QModelIndex &index, int role) const
     case ProposalRole:
         return rec->name;
     case AmountRole:
-        return rec->amount;
+        return (long long) rec->amount;
     case StartDateRole:
         return rec->start_epoch;
     case EndDateRole:
         return rec->end_epoch;
     case YesVotesRole:
-        return rec->yesVotes;
+        return (long long) rec->yesVotes;
     case NoVotesRole:
-        return rec->noVotes;
+        return (long long) rec->noVotes;
     case AbsoluteYesVotesRole:
-        return rec->absoluteYesVotes;
+        return (long long) rec->absoluteYesVotes;
     case PercentageRole:
-        return rec->percentage;
+        return (long long) rec->percentage;
     case ProposalUrlRole:
         return rec->url;
     case ProposalHashRole:
